fix: address PR #52 review feedback — CLI output, semver, polling, installer
- Fix inverted condition in cli.ts that showed raw JSON on TTY instead of
  human-friendly output (!== → ===)
- Replace lexicographic version comparison with proper semver segment
  comparison in update-check.ts
- Extract shared pollForResponse to src/lib/poll.ts with sequence tracking
  and ElnoraError (deduplicates create.ts/send.ts)
- Add [Environment]::UserInteractive guard to install.ps1 Read-Host to
  prevent hanging in non-interactive contexts
- Remove dead if (!data.streamed) branch in tasks create formatOutput
- Add --stream to projects/create.ts suggested command for consistency
