+++ src/commands/tasks/create.ts
@@ -1,6 +1,7 @@
 import { z } from "zod";
 import type { ElnoraCommand } from "../../core/command.js";
 import type { OutputFormat } from "../../lib/output.js";
+import { pollForResponse } from "../../lib/poll.js";
 import { resolveApiKey } from "../../lib/profiles.js";
 import { collectStreamResponse, streamTask } from "../../lib/stream.js";
 import { StreamRenderer } from "../../lib/stream-renderer.js";
@@ -14,43 +15,6 @@ const inputSchema = z.object({
 });
 
 type Input = z.infer<typeof inputSchema>;
-
-/**
- * Poll GET /tasks/{id}/messages until an assistant message appears.
- */
-async function pollForResponse(
-	client: {
-		get: (
-			endpoint: string,
-			opts?: { pathParams?: Record<string, string>; queryParams?: Record<string, string | number> },
-		) => Promise<unknown>;
-	},
-	taskId: string,
-	timeoutMs = 120_000,
-): Promise<unknown> {
-	const deadline = Date.now() + timeoutMs;
-	while (Date.now() < deadline) {
-		const messages = await client.get("task_messages", {
-			pathParams: { id: taskId },
-			queryParams: { limit: 10 },
-		});
-
-		const items = Array.isArray(messages)
-			? messages
-			: (((messages as Record<string, unknown>)?.items as unknown[]) ?? []);
-		if (Array.isArray(items)) {
-			const assistantMsg = items.find((m: unknown) => {
-				const msg = m as Record<string, unknown>;
-				return msg.role === "assistant" || msg.role === "Assistant";
-			});
-			if (assistantMsg) return assistantMsg;
-		}
-
-		await new Promise((r) => setTimeout(r, 2000));
-	}
-
-	throw new Error(`Timed out waiting for agent response (120s). Check with: elnora tasks messages ${taskId}`);
-}
 
 export const tasksCreate: ElnoraCommand<Input> = {
 	name: "tasks.create",
@@ -104,10 +68,12 @@ export const tasksCreate: ElnoraCommand<Input> = {
 		if (format === "compact") return data.id ?? JSON.stringify(output);
 		if (data.streamed) return ""; // stream already rendered to stdout
 
-		const lines = [`✓ Task created: ${data.id}`];
-		if (!data.streamed) {
-			lines.push("", "Send a message:", `  elnora tasks send ${data.id} --message "Your message" --stream`);
-		}
+		const lines = [
+			`✓ Task created: ${data.id}`,
+			"",
+			"Send a message:",
+			`  elnora tasks send ${data.id} --message "Your message" --stream`,
+		];
 		return lines.join("\n");
 	},
 };
