+++ src/commands/tasks/send.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 import type { ElnoraCommand } from "../../core/command.js";
-import { ElnoraError } from "../../lib/errors.js";
 import type { OutputFormat } from "../../lib/output.js";
+import { pollForResponse } from "../../lib/poll.js";
 import { resolveApiKey } from "../../lib/profiles.js";
 import { collectStreamResponse, streamTask } from "../../lib/stream.js";
 import { StreamRenderer } from "../../lib/stream-renderer.js";
@@ -17,51 +17,6 @@ const inputSchema = z.object({
 });
 
 type Input = z.infer<typeof inputSchema>;
-
-/**
- * Poll GET /tasks/{id}/messages until an assistant message appears
- * with a sequence number higher than the user's message.
- */
-async function pollForResponse(
-	client: {
-		get: (
-			endpoint: string,
-			opts?: { pathParams?: Record<string, string>; queryParams?: Record<string, string | number> },
-		) => Promise<unknown>;
-	},
-	taskId: string,
-	userMessageSequence: number,
-	timeoutMs = 120_000,
-): Promise<unknown> {
-	const deadline = Date.now() + timeoutMs;
-	while (Date.now() < deadline) {
-		const messages = await client.get("task_messages", {
-			pathParams: { id: taskId },
-			queryParams: { limit: 10 },
-		});
-
-		const items = Array.isArray(messages)
-			? messages
-			: (((messages as Record<string, unknown>)?.items as unknown[]) ?? []);
-		if (Array.isArray(items)) {
-			const assistantMsg = items.find((m: unknown) => {
-				const msg = m as Record<string, unknown>;
-				return (
-					(msg.role === "assistant" || msg.role === "Assistant") &&
-					((msg.sequence as number) ?? (msg.sequenceNumber as number) ?? 0) > userMessageSequence
-				);
-			});
-			if (assistantMsg) return assistantMsg;
-		}
-
-		await new Promise((r) => setTimeout(r, 2000));
-	}
-
-	throw new ElnoraError("Timed out waiting for agent response (120s)", {
-		code: "RESPONSE_TIMEOUT",
-		suggestion: `The agent may still be processing. Check with: elnora tasks messages ${taskId}`,
-	});
-}
 
 export const tasksSend: ElnoraCommand<Input> = {
 	name: "tasks.send",
